fixed enum treatiment logic
